requirements.txt
google-adk>=0.4.0
google-genai>=0.8.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
google-cloud-firestore>=2.16.0
a2a-sdk>=0.3.0